Simple Registration form C#
A simple registarion form like the one you've seen in a place where you need to be registered to enter.

How to use it?

Simple just Fill what it will ask you and tell it where to export the text file and finish.

see how simple it is.

The project is programmed in C# programming language feel free to use as you want and even make it better